Separate out TypeStructure
